// frontend/src/api/model/professionalCreate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Hospital Internment API
 * OpenAPI spec version: 0.1.0
 */
import type { ProfessionalSpecialtyCreate } from './professionalSpecialtyCreate';

export interface ProfessionalCreate {
  first_name: string;
  last_name: string;
  document_type: string;
  document_number: string;
  email: string | null;
  phone: string | null;
  specialties: ProfessionalSpecialtyCreate[];
}